Add slab allocator with Vale-style random generational references
Implement a custom slab allocator for heap objects with 64-bit random
generations for weak reference validation:

- Platform abstraction layer (vmem.hpp, vmem_win32.cpp, vmem_unix.cpp)
  for reserve/commit/decommit/remap_to_zero operations
- Slab allocator with 8 size classes (32B-4KB) plus large object support
- Tombstoning: freed objects keep memory mapped for safe weak ref checks
- 64-bit random generations via xorshift128+ PRNG (no wrap-around issues)
- ObjectHeader expanded to 24 bytes with flags (FLAG_ALIVE/FLAG_TOMBSTONE)
- Updated WEAK_CHECK opcode to perform full generation validation

Key changes:
- ObjectHeader: u32 -> u64 weak_generation, added flags field
- Value: as_weak.generation u32 -> u64
- RoxyVM: added SlabAllocator* allocator member
- object_alloc/object_free: use SlabAllocator instead of malloc/free
- weak_ref_valid: checks is_alive() && generation match

Tests: 19 new tests (8 unit + 11 stress) with 244K+ assertions covering
allocation patterns, generation uniqueness, tombstoning, and weak ref
validation under heavy churn.

+++ include/roxy/vm/value.hpp
@@ -28,7 +28,7 @@ struct Value {
         void* as_ptr;
         struct {
             void* ptr;
-            u32 generation;
+            u64 generation;  // 64-bit random generation for weak reference validation
         } as_weak;
     };
 
@@ -71,7 +71,7 @@ struct Value {
         return v;
     }
 
-    static Value make_weak(void* p, u32 generation) {
+    static Value make_weak(void* p, u64 generation) {
         Value v;
         v.type = Weak;
         v.as_weak.ptr = p;
